Switch to App Router (everything except API moved from pages)
