+++ src/app/Dashboard/ViewReport/page.jsx
@@ -18,9 +18,8 @@ import {
   VStack,
 } from "@chakra-ui/react";
 import urls from "utils/urls";
-import { updateReport, getReport } from "src/actions/Report";
+import { updateReport, getReport } from "src/app/actions/Report";
 import { getSession } from "next-auth/react";
-import Layout from "src/app/Index/layout";
 import Report from "../NewReport/report";
 import ReportPreview from "./reportPreview";
 
@@ -145,7 +144,3 @@ export default function Page() {
     </>
   );
 }
-
-Page.getLayout = function (page) {
-  return <Layout>{page}</Layout>;
-};
