+++ src/app/layout.js
@@ -1,6 +1,6 @@
 export default function RootLayout({ children }) {
   return (
-    <html lang="en">
+    <html>
       <body>{children}</body>
     </html>
   );
