+++ src/middleware.js
@@ -3,10 +3,11 @@ export { default } from "next-auth/middleware";
 export const config = {
   matcher: [
     "/",
-    "/guidelines",
-    "/profile",
-    "/reports",
-    "/newreport",
-    "/viewreport",
+    "/Index",
+    "/Guidelines",
+    "/Profile",
+    "/Reports",
+    "/NewReport",
+    "/ViewReport",
   ],
 };
