--- src/pages/_app.jsx
@@ -1,17 +1,0 @@
-import { ChakraProvider } from "@chakra-ui/react";
-import { SessionProvider } from "next-auth/react";
-import { customTheme } from "../app/styles/customTheme";
-
-function MyApp({ Component, pageProps: { session, ...pageProps } }) {
-  const renderWithLayout = Component.getLayout || ((page) => page);
-
-  return (
-    <SessionProvider session={session}>
-      <ChakraProvider theme={customTheme}>
-        {renderWithLayout(<Component {...pageProps} />)}
-      </ChakraProvider>
-    </SessionProvider>
-  );
-}
-
-export default MyApp;
