--- src/pages/guidelines.js
@@ -1,3 +1,0 @@
-import Guidelines from "../app/Guidelines/page";
-
-export default Guidelines;
