--- src/pages/index.js
@@ -1,3 +1,0 @@
-import HomePage from "../app/Index/page";
-
-export default HomePage;
