--- src/pages/login.js
@@ -1,3 +1,0 @@
-import Login from "../app/Login/page";
-
-export default Login;
