--- src/pages/logout.js
@@ -1,3 +1,0 @@
-import Logout from "../app/Logout/page";
-
-export default Logout;
