--- src/pages/newreport.js
@@ -1,3 +1,0 @@
-import NewReport from "../app/NewReport/page";
-
-export default NewReport;
