--- src/pages/profile.js
@@ -1,3 +1,0 @@
-import Profile from "../app/Profile/page";
-
-export default Profile;
