--- src/pages/reports.js
@@ -1,3 +1,0 @@
-import Reports from "../app/Reports/page";
-
-export default Reports;
