--- src/pages/viewreport.js
@@ -1,3 +1,0 @@
-import ViewReport from "../app/ViewReport/page";
-
-export default ViewReport;
