+++ utils/urls.js
@@ -1,27 +1,13 @@
 function getBaseURL() {
   // if backend
   if (process.env.VERCEL_URL) {
-    return `https://${process.env.VERCEL_URL}`;
-  }
-  // if client-side
-  if (process.env.NEXT_PUBLIC_VERCEL_URL) {
-    return `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`;
+    return `https://${process.env.NEXTAUTH_URL}`;
   }
   return "http://localhost:3000";
 }
 
 export default {
   baseUrl: getBaseURL(),
-  pages: {
-    home: "/",
-    guidelines: "/guidelines",
-    reports: "/reports",
-    profile: "/profile",
-    login: "/login",
-    newreport: "/newreport",
-    viewreport: "/newreport",
-    logout: "/logout",
-  },
   api: {
     reports: {
       create: "/api/reports/create",
